pyproject.toml
[project]
name = "voltis"
version = "0.1.0"
description = "Library for ebooks and comics (& co)"
requires-python = ">=3.14"
dependencies = [
    "anyio>=4.11.0",
    "bcrypt>=5.0.0",
    "click>=8.3.0",
    "fastapi[standard]>=0.118.3",
    "httpx>=0.28.1",
    "orjson>=3.11.3",
    "psycopg[binary]>=3.2.10",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.44",
    "structlog>=25.4.0",
    "uvicorn>=0.37.0",
]

[project.scripts]
voltis = "voltis.cli.main:main"

[tool.uv]
package = true
config-settings = { editable_mode = "compat" }

[tool.ruff]
line-length = 100

[tool.pyright]
typeCheckingMode = "standard"
# extraPaths = ["src"]
include = ["src"]

[dependency-groups]
dev = ["pyright>=1.1.406", "pytest>=8.4.2", "ruff>=0.14.0"]
